Time: 6 ms (96.37%), Space: 11.5 MB (50.27%) - LeetHub

+++ 852-peak-index-in-a-mountain-array/852-peak-index-in-a-mountain-array.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     int peakIndexInMountainArray(vector<int>& arr) {
-        long maxi = *max_element(arr.begin(),arr.end());
+        int maxi = *max_element(arr.begin(),arr.end());
         return find(arr.begin(),arr.end(),maxi)-arr.begin();
     }
 };
